+++ core-strato/core-api/app/Main.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell   #-}
 {-# LANGUAGE TypeOperators     #-}
 
 
@@ -10,12 +11,17 @@
 
 module Main where
 
+import           Blockchain.Output
 import           Control.Lens.Operators
-import           Control.Monad.Logger
 import           Data.Aeson
+import qualified Data.ByteString.Lazy.Char8      as BLC
+import qualified Data.HashMap.Strict.InsOrd           as H
 import           Data.Proxy
 import           Data.Swagger
 import           Database.Persist.Postgresql
+import           HFlags
+import           Network.HTTP.Types.Status
+import           Network.Wai
 import           Network.Wai.Handler.Warp
 import           Network.Wai.Middleware.Cors
 import           Network.Wai.Middleware.Prometheus
@@ -25,8 +31,13 @@ import           Servant.Multipart
 import           Servant.Swagger
 import           Servant.Swagger.UI
 
+
+
+
 import           BlockApps.Init
 import           Blockchain.EthConf
+
+import           Text.Tools
 
 import qualified Handlers.AccountInfo            as Account
 import qualified Handlers.BatchTransactionResult as BatchTransactionResult
@@ -98,6 +109,7 @@ coreAPI = Proxy
 
 main :: IO ()
 main = do
+  _ <- $initHFlags "Core API"
   let theDoc = toSwagger (Proxy :: Proxy CoreAPI)
                & info.title .~ "Strato API"
                & info.description ?~
@@ -116,10 +128,18 @@ app pool theDoc =
   $ instrumentApp "core-api"
   $ logStdoutDev
   $ cors (const $ Just simpleCorsResourcePolicy{corsRequestHeaders=["Content-Type"]})
-  $ serve (Proxy :: Proxy (CoreAPI :<|> SwaggerSchemaUI "swagger-ui" "swagger.json")) $ (coreServer pool :<|> swaggerSchemaUIServer theDoc)
+--  $ serve (Proxy :: Proxy (CoreAPI :<|> SwaggerSchemaUI "swagger-ui" "swagger.json")) $ (coreServer pool :<|> swaggerSchemaUIServer theDoc)
+  $ serve (Proxy :: Proxy (CoreAPI :<|> SwaggerSchemaUI "swagger-ui" "swagger.json" :<|> Raw)) $ (coreServer pool :<|> swaggerSchemaUIServer theDoc :<|> Tagged serveCustom404)
 
 
 
+serveCustom404 :: Application
+serveCustom404 x respond =
+  respond $ responseLBS notFound404 [("Content-Type", "text/plain")] $ BLC.pack
+  $ "There is no content at: " ++ show (rawPathInfo x)
+  ++ "\nHere are the available routes:" ++ tab ("\n" ++ unlines allPaths) ++ "\n"
+  where
+    allPaths = H.keys $ _swaggerPaths $ toSwagger (Proxy :: Proxy CoreAPI)
 
 
 ----------