+++ core-strato/core-api/src/Handlers/Chain.hs
@@ -8,7 +8,6 @@ module Handlers.Chain (
   ) where
 
 import           Control.Monad
-import           Control.Monad.Logger
 import           Control.Monad.IO.Class
 import qualified Data.ByteString.Char8          as BC
 import qualified Data.ByteString.Lazy.Char8     as BLC
@@ -27,11 +26,13 @@ import           Blockchain.Data.ChainInfoDB
 import           Blockchain.Data.TXOrigin
 import           Blockchain.EthConf             (runKafkaConfigured)
 import           Blockchain.ExtWord
+import           Blockchain.Output
 import           Blockchain.Sequencer.Event     (IngestEvent (IEGenesis), IngestGenesis (..))
 import           Blockchain.Sequencer.Kafka     (writeUnseqEvents)
 import           Blockchain.Strato.Model.ChainId
 import           Blockchain.Strato.Model.CodePtr
 import           Blockchain.Strato.Model.SHA
+import           Blockchain.Strato.Model.Util
 import           Blockchain.TypeLits
 import           SQLM
 
@@ -56,23 +57,23 @@ getChain pool mChainId = liftIO $ runSQLM pool $
     Just (ChainId chainid) -> getChainInfos [chainid]
     
 postChain :: ChainInfo -> Handler Text
-postChain ci = runStdoutLoggingT $ do
+postChain ci = runLoggingT $ do
     case processChainInfos [ci] of
       Left (_, err) -> throwError $ err400{ errBody=BLC.pack $ "invalid args: " ++ err }
       Right [] -> error "postChainR: The impossible happened. processChainInfos succeeded, but returned an empty list"
       Right (cid:_) -> do
-        let hexCid = T.pack $ showHex cid ""
+        let hexCid = T.pack $ padZeros 64 $ showHex cid ""
         $logDebugS "postChainR" . T.pack $ show ci
         $logInfoS "postChainR" hexCid
         emitKafkaTransactions $ [(cid,ci)]
         return hexCid
 
 postChains :: [ChainInfo] -> Handler [Text]
-postChains cis = runStdoutLoggingT $ do
+postChains cis = runLoggingT $ do
   case processChainInfos cis of
       Left (i, err) -> throwError err400{ errBody=BLC.pack $ "invalid args at index " ++ show i ++ ": " ++ err }
       Right cids -> do
-        let hexCids = map (T.pack . flip showHex "") cids
+        let hexCids = map (T.pack . padZeros 64 . flip showHex "") cids
         $logDebugS "postChainsR" . T.pack $ show cis
         $logInfoS "postChainsR" $ T.intercalate ", " hexCids
         emitKafkaTransactions $ zip cids cis