+++ blockapps-rest/ReleaseNotes.md
@@ -1,6 +1,8 @@
 ## RELEASE NOTES
 
 ### Version 7.6.6
+
+- Added /transaction/parallel endpoint
 - Bug fix for createContractList
 
 ### Version 7.6.5